tests/wpt/tests/html/semantics/embedded-content/media-elements/autoplay-supported-by-permissions-policy.html

<!DOCTYPE html>
<title>Test that autoplay is advertised in the feature list</title>
<link rel="help" href="https://w3c.github.io/webappsec-permissions-policy/#dom-permissionspolicy-features">
<link rel="help" href="https://html.spec.whatwg.org/multipage/infrastructure.html#policy-controlled-features">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => {
    assert_in_array('autoplay', document.permissionsPolicy.features());
}, 'document.permissionsPolicy.features should advertise autoplay.');
</script>
